+++ src/game/Object/ObjectMgr.cpp
@@ -2154,293 +2154,6 @@ uint32 ObjectMgr::GetTaxiMountDisplayId(uint32 id, Team team, bool allowed_alt_t
 }
 
 /**
- * @brief Loads graveyard to zone links from the database.
- */
-void ObjectMgr::LoadGraveyardZones()
-{
-    mGraveYardMap.clear();                                  // need for reload case
-
-    QueryResult* result = WorldDatabase.Query("SELECT `id`,`ghost_zone`,`faction` FROM `game_graveyard_zone`");
-
-    uint32 count = 0;
-
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outString(">> Loaded %u graveyard-zone links", count);
-        sLog.outString();
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        ++count;
-        bar.step();
-
-        Field* fields = result->Fetch();
-
-        uint32 safeLocId = fields[0].GetUInt32();
-        uint32 zoneId = fields[1].GetUInt32();
-        uint32 team   = fields[2].GetUInt32();
-
-        WorldSafeLocsEntry const* entry = sWorldSafeLocsStore.LookupEntry(safeLocId);
-        if (!entry)
-        {
-            sLog.outErrorDb("Table `game_graveyard_zone` has record for not existing graveyard (WorldSafeLocs.dbc id) %u, skipped.", safeLocId);
-            continue;
-        }
-
-        AreaTableEntry const* areaEntry = GetAreaEntryByAreaID(zoneId);
-        if (!areaEntry)
-        {
-            sLog.outErrorDb("Table `game_graveyard_zone` has record for not existing zone id (%u), skipped.", zoneId);
-            continue;
-        }
-
-        if (areaEntry->zone != 0)
-        {
-            sLog.outErrorDb("Table `game_graveyard_zone` has record subzone id (%u) instead of zone, skipped.", zoneId);
-            continue;
-        }
-
-        if (team != TEAM_BOTH_ALLOWED && team != HORDE && team != ALLIANCE)
-        {
-            sLog.outErrorDb("Table `game_graveyard_zone` has record for non player faction (%u), skipped.", team);
-            continue;
-        }
-
-        if (!AddGraveYardLink(safeLocId, zoneId, Team(team), false))
-        {
-            sLog.outErrorDb("Table `game_graveyard_zone` has a duplicate record for Graveyard (ID: %u) and Zone (ID: %u), skipped.", safeLocId, zoneId);
-        }
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString(">> Loaded %u graveyard-zone links", count);
-    sLog.outString();
-}
-
-/**
- * @brief Gets the closest suitable graveyard for a location and faction.
- *
- * @param x The world x coordinate.
- * @param y The world y coordinate.
- * @param z The world z coordinate.
- * @param MapId The map id.
- * @param team The player's faction.
- * @return The closest matching graveyard safe location, or null if none is linked.
- */
-WorldSafeLocsEntry const* ObjectMgr::GetClosestGraveYard(float x, float y, float z, uint32 MapId, Team team)
-{
-    // search for zone associated closest graveyard
-    uint32 zoneId = sTerrainMgr.GetZoneId(MapId, x, y, z);
-
-    // Simulate std. algorithm:
-    //   found some graveyard associated to (ghost_zone,ghost_map)
-    //
-    //   if mapId == graveyard.mapId (ghost in plain zone or city or battleground) and search graveyard at same map
-    //     then check faction
-    //   if mapId != graveyard.mapId (ghost in instance) and search any graveyard associated
-    //     then check faction
-    GraveYardMapBounds bounds = mGraveYardMap.equal_range(zoneId);
-
-    if (bounds.first == bounds.second)
-    {
-        sLog.outErrorDb("Table `game_graveyard_zone` incomplete: Zone %u Team %u does not have a linked graveyard.", zoneId, uint32(team));
-        return NULL;
-    }
-
-    // at corpse map
-    bool foundNear = false;
-    float distNear;
-    WorldSafeLocsEntry const* entryNear = NULL;
-
-    // at entrance map for corpse map
-    bool foundEntr = false;
-    float distEntr;
-    WorldSafeLocsEntry const* entryEntr = NULL;
-
-    // some where other
-    WorldSafeLocsEntry const* entryFar = NULL;
-
-    MapEntry const* mapEntry = sMapStore.LookupEntry(MapId);
-
-    for (GraveYardMap::const_iterator itr = bounds.first; itr != bounds.second; ++itr)
-    {
-        GraveYardData const& data = itr->second;
-
-        // Checked on load
-        WorldSafeLocsEntry const* entry = sWorldSafeLocsStore.LookupEntry(data.safeLocId);
-
-        // skip enemy faction graveyard
-        // team == TEAM_BOTH_ALLOWED case can be at call from .neargrave
-        // TEAM_INVALID != team for all teams
-        if (data.team != TEAM_BOTH_ALLOWED && data.team != team && team != TEAM_BOTH_ALLOWED)
-        {
-            continue;
-        }
-
-        // find now nearest graveyard at other (continent) map
-        if (MapId != entry->map_id)
-        {
-            // if find graveyard at different map from where entrance placed (or no entrance data), use any first
-            if (!mapEntry ||
-                    mapEntry->ghost_entrance_map < 0 ||
-                    uint32(mapEntry->ghost_entrance_map) != entry->map_id ||
-                    (mapEntry->ghost_entrance_x == 0 && mapEntry->ghost_entrance_y == 0))
-            {
-                // not have any coordinates for check distance anyway
-                entryFar = entry;
-                continue;
-            }
-
-            // at entrance map calculate distance (2D);
-            float dist2 = (entry->x - mapEntry->ghost_entrance_x) * (entry->x - mapEntry->ghost_entrance_x)
-                          + (entry->y - mapEntry->ghost_entrance_y) * (entry->y - mapEntry->ghost_entrance_y);
-            if (foundEntr)
-            {
-                if (dist2 < distEntr)
-                {
-                    distEntr = dist2;
-                    entryEntr = entry;
-                }
-            }
-            else
-            {
-                foundEntr = true;
-                distEntr = dist2;
-                entryEntr = entry;
-            }
-        }
-        // find now nearest graveyard at same map
-        else
-        {
-            float dist2 = (entry->x - x) * (entry->x - x) + (entry->y - y) * (entry->y - y) + (entry->z - z) * (entry->z - z);
-            if (foundNear)
-            {
-                if (dist2 < distNear)
-                {
-                    distNear = dist2;
-                    entryNear = entry;
-                }
-            }
-            else
-            {
-                foundNear = true;
-                distNear = dist2;
-                entryNear = entry;
-            }
-        }
-    }
-
-    if (entryNear)
-    {
-        return entryNear;
-    }
-
-    if (entryEntr)
-    {
-        return entryEntr;
-    }
-
-    return entryFar;
-}
-
-/**
- * @brief Finds graveyard link data for a graveyard id and zone.
- *
- * @param id The graveyard safe location id.
- * @param zoneId The zone id.
- * @return The graveyard link data, or null if not found.
- */
-GraveYardData const* ObjectMgr::FindGraveYardData(uint32 id, uint32 zoneId) const
-{
-    GraveYardMapBounds bounds = mGraveYardMap.equal_range(zoneId);
-
-    for (GraveYardMap::const_iterator itr = bounds.first; itr != bounds.second; ++itr)
-    {
-        if (itr->second.safeLocId == id)
-        {
-            return &itr->second;
-        }
-    }
-
-    return NULL;
-}
-
-/**
- * @brief Adds a graveyard link for a zone and optional database persistence.
- *
- * @param id The graveyard safe location id.
- * @param zoneId The zone id.
- * @param team The faction restriction for the link.
- * @param inDB true to also insert the link into the database.
- * @return true if the link was added; otherwise, false.
- */
-bool ObjectMgr::AddGraveYardLink(uint32 id, uint32 zoneId, Team team, bool inDB)
-{
-    if (FindGraveYardData(id, zoneId))                      // This ensures that (safeLoc)Id,  zoneId is unique in mGraveYardMap
-    {
-        return false;
-    }
-
-    // add link to loaded data
-    GraveYardData data;
-    data.safeLocId = id;
-    data.team = team;
-
-    mGraveYardMap.insert(GraveYardMap::value_type(zoneId, data));
-
-    // add link to DB
-    if (inDB)
-    {
-        WorldDatabase.PExecuteLog("INSERT INTO `game_graveyard_zone` (`id`,`ghost_zone`,`faction`) VALUES ('%u', '%u','%u')", id, zoneId, uint32(team));
-    }
-
-    return true;
-}
-
-/**
- * @brief Updates the faction restriction for an existing graveyard link.
- *
- * @param id The graveyard safe location id.
- * @param zoneId The zone id.
- * @param team The faction restriction to set.
- */
-void ObjectMgr::SetGraveYardLinkTeam(uint32 id, uint32 zoneId, Team team)
-{
-    std::pair<GraveYardMap::iterator, GraveYardMap::iterator> bounds = mGraveYardMap.equal_range(zoneId);
-
-    for (GraveYardMap::iterator itr = bounds.first; itr != bounds.second; ++itr)
-    {
-        GraveYardData& data = itr->second;
-
-        // skip not matching safezone id
-        if (data.safeLocId != id)
-        {
-            continue;
-        }
-
-        data.team = team;                                   // Validate link
-        return;
-    }
-
-    if (team == TEAM_INVALID)
-    {
-        return;
-    }
-
-    // Link expected but not exist.
-    sLog.outErrorDb("ObjectMgr::SetGraveYardLinkTeam called for safeLoc %u, zoneId %u, but no graveyard link for this found in database.", id, zoneId);
-    AddGraveYardLink(id, zoneId, team);                     // Add to prevent further error message and correct mechanismn
-}
-
-/**
  * @brief Loads area trigger teleport destinations and access requirements.
  */
 void ObjectMgr::LoadAreaTriggerTeleports()
@@ -3454,279 +3167,6 @@ void ObjectMgr::LoadCorpses()
     delete result;
 
     sLog.outString(">> Loaded %u corpses", count);
-    sLog.outString();
-}
-
-/**
- * @brief Loads faction-specific reputation reward rate multipliers.
- */
-void ObjectMgr::LoadReputationRewardRate()
-{
-    m_RepRewardRateMap.clear();                             // for reload case
-
-    uint32 count = 0;
-    QueryResult* result = WorldDatabase.Query("SELECT `faction`, `quest_rate`, `creature_rate`, `spell_rate` FROM `reputation_reward_rate`");
-
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outErrorDb(">> Loaded `reputation_reward_rate`, table is empty!");
-        sLog.outString();
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        bar.step();
-
-        Field* fields = result->Fetch();
-
-        uint32 factionId        = fields[0].GetUInt32();
-
-        RepRewardRate repRate;
-
-        repRate.quest_rate      = fields[1].GetFloat();
-        repRate.creature_rate   = fields[2].GetFloat();
-        repRate.spell_rate      = fields[3].GetFloat();
-
-        FactionEntry const* factionEntry = sFactionStore.LookupEntry(factionId);
-        if (!factionEntry)
-        {
-            sLog.outErrorDb("Faction (faction.dbc) %u does not exist but is used in `reputation_reward_rate`", factionId);
-            continue;
-        }
-
-        if (repRate.quest_rate < 0.0f)
-        {
-            sLog.outErrorDb("Table reputation_reward_rate has quest_rate with invalid rate %f, skipping data for faction %u", repRate.quest_rate, factionId);
-            continue;
-        }
-
-        if (repRate.creature_rate < 0.0f)
-        {
-            sLog.outErrorDb("Table reputation_reward_rate has creature_rate with invalid rate %f, skipping data for faction %u", repRate.creature_rate, factionId);
-            continue;
-        }
-
-        if (repRate.spell_rate < 0.0f)
-        {
-            sLog.outErrorDb("Table reputation_reward_rate has spell_rate with invalid rate %f, skipping data for faction %u", repRate.spell_rate, factionId);
-            continue;
-        }
-
-        m_RepRewardRateMap[factionId] = repRate;
-
-        ++count;
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString(">> Loaded %u reputation_reward_rate", count);
-    sLog.outString();
-}
-
-/**
- * @brief Loads creature reputation rewards granted on kill.
- */
-void ObjectMgr::LoadReputationOnKill()
-{
-    uint32 count = 0;
-
-    //                                                0            1                     2
-    QueryResult* result = WorldDatabase.Query("SELECT `creature_id`, `RewOnKillRepFaction1`, `RewOnKillRepFaction2`,"
-                          //   3             4             5                   6             7             8                   9
-                          "`IsTeamAward1`, `MaxStanding1`, `RewOnKillRepValue1`, `IsTeamAward2`, `MaxStanding2`, `RewOnKillRepValue2`, `TeamDependent` "
-                          "FROM `creature_onkill_reputation`");
-
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outErrorDb(">> Loaded 0 creature award reputation definitions. DB table `creature_onkill_reputation` is empty.");
-        sLog.outString();
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        Field* fields = result->Fetch();
-        bar.step();
-
-        uint32 creature_id = fields[0].GetUInt32();
-
-        ReputationOnKillEntry repOnKill;
-        repOnKill.repfaction1          = fields[1].GetUInt32();
-        repOnKill.repfaction2          = fields[2].GetUInt32();
-        repOnKill.is_teamaward1        = fields[3].GetBool();
-        repOnKill.reputation_max_cap1  = fields[4].GetUInt32();
-        repOnKill.repvalue1            = fields[5].GetInt32();
-        repOnKill.is_teamaward2        = fields[6].GetBool();
-        repOnKill.reputation_max_cap2  = fields[7].GetUInt32();
-        repOnKill.repvalue2            = fields[8].GetInt32();
-        repOnKill.team_dependent       = fields[9].GetUInt8();
-
-        if (!GetCreatureTemplate(creature_id))
-        {
-            sLog.outErrorDb("Table `creature_onkill_reputation` have data for nonexistent creature entry (%u), skipped", creature_id);
-            continue;
-        }
-
-        if (repOnKill.repfaction1)
-        {
-            FactionEntry const* factionEntry1 = sFactionStore.LookupEntry(repOnKill.repfaction1);
-            if (!factionEntry1)
-            {
-                sLog.outErrorDb("Faction (faction.dbc) %u does not exist but is used in `creature_onkill_reputation`", repOnKill.repfaction1);
-                continue;
-            }
-        }
-
-        if (repOnKill.repfaction2)
-        {
-            FactionEntry const* factionEntry2 = sFactionStore.LookupEntry(repOnKill.repfaction2);
-            if (!factionEntry2)
-            {
-                sLog.outErrorDb("Faction (faction.dbc) %u does not exist but is used in `creature_onkill_reputation`", repOnKill.repfaction2);
-                continue;
-            }
-        }
-
-        mRepOnKill[creature_id] = repOnKill;
-
-        ++count;
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString(">> Loaded %u creature award reputation definitions", count);
-    sLog.outString();
-}
-
-/**
- * @brief Loads reputation spillover templates for shared faction gains.
- */
-void ObjectMgr::LoadReputationSpilloverTemplate()
-{
-    m_RepSpilloverTemplateMap.clear();                      // for reload case
-
-    uint32 count = 0;
-    QueryResult* result = WorldDatabase.Query("SELECT `faction`, `faction1`, `rate_1`, `rank_1`, `faction2`, `rate_2`, `rank_2`, `faction3`, `rate_3`, `rank_3`, `faction4`, `rate_4`, `rank_4` FROM `reputation_spillover_template`");
-
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outString(">> Loaded `reputation_spillover_template`, table is empty.");
-        sLog.outString();
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        bar.step();
-
-        Field* fields = result->Fetch();
-
-        uint32 factionId                = fields[0].GetUInt32();
-
-        RepSpilloverTemplate repTemplate;
-
-        repTemplate.faction[0]          = fields[1].GetUInt32();
-        repTemplate.faction_rate[0]     = fields[2].GetFloat();
-        repTemplate.faction_rank[0]     = fields[3].GetUInt32();
-        repTemplate.faction[1]          = fields[4].GetUInt32();
-        repTemplate.faction_rate[1]     = fields[5].GetFloat();
-        repTemplate.faction_rank[1]     = fields[6].GetUInt32();
-        repTemplate.faction[2]          = fields[7].GetUInt32();
-        repTemplate.faction_rate[2]     = fields[8].GetFloat();
-        repTemplate.faction_rank[2]     = fields[9].GetUInt32();
-        repTemplate.faction[3]          = fields[10].GetUInt32();
-        repTemplate.faction_rate[3]     = fields[11].GetFloat();
-        repTemplate.faction_rank[3]     = fields[12].GetUInt32();
-
-        FactionEntry const* factionEntry = sFactionStore.LookupEntry(factionId);
-
-        if (!factionEntry)
-        {
-            sLog.outErrorDb("Faction (faction.dbc) %u does not exist but is used in `reputation_spillover_template`", factionId);
-            continue;
-        }
-
-        if (factionEntry->team == 0)
-        {
-            sLog.outErrorDb("Faction (faction.dbc) %u in `reputation_spillover_template` does not belong to any team, skipping", factionId);
-            continue;
-        }
-
-        for (uint32 i = 0; i < MAX_SPILLOVER_FACTIONS; ++i)
-        {
-            if (repTemplate.faction[i])
-            {
-                FactionEntry const* factionSpillover = sFactionStore.LookupEntry(repTemplate.faction[i]);
-
-                if (!factionSpillover)
-                {
-                    sLog.outErrorDb("Spillover faction (faction.dbc) %u does not exist but is used in `reputation_spillover_template` for faction %u, skipping", repTemplate.faction[i], factionId);
-                    continue;
-                }
-
-                if (factionSpillover->reputationListID < 0)
-                {
-                    sLog.outErrorDb("Spillover faction (faction.dbc) %u for faction %u in `reputation_spillover_template` can not be listed for client, and then useless, skipping", repTemplate.faction[i], factionId);
-                    continue;
-                }
-
-                if (repTemplate.faction_rank[i] >= MAX_REPUTATION_RANK)
-                {
-                    sLog.outErrorDb("Rank %u used in `reputation_spillover_template` for spillover faction %u is not valid, skipping", repTemplate.faction_rank[i], repTemplate.faction[i]);
-                    continue;
-                }
-            }
-        }
-
-        FactionEntry const* factionEntry0 = sFactionStore.LookupEntry(repTemplate.faction[0]);
-        if (repTemplate.faction[0] && !factionEntry0)
-        {
-            sLog.outErrorDb("Faction (faction.dbc) %u does not exist but is used in `reputation_spillover_template`", repTemplate.faction[0]);
-            continue;
-        }
-        FactionEntry const* factionEntry1 = sFactionStore.LookupEntry(repTemplate.faction[1]);
-        if (repTemplate.faction[1] && !factionEntry1)
-        {
-            sLog.outErrorDb("Faction (faction.dbc) %u does not exist but is used in `reputation_spillover_template`", repTemplate.faction[1]);
-            continue;
-        }
-        FactionEntry const* factionEntry2 = sFactionStore.LookupEntry(repTemplate.faction[2]);
-        if (repTemplate.faction[2] && !factionEntry2)
-        {
-            sLog.outErrorDb("Faction (faction.dbc) %u does not exist but is used in `reputation_spillover_template`", repTemplate.faction[2]);
-            continue;
-        }
-        FactionEntry const* factionEntry3 = sFactionStore.LookupEntry(repTemplate.faction[3]);
-        if (repTemplate.faction[3] && !factionEntry3)
-        {
-            sLog.outErrorDb("Faction (faction.dbc) %u does not exist but is used in `reputation_spillover_template`", repTemplate.faction[3]);
-            continue;
-        }
-
-        m_RepSpilloverTemplateMap[factionId] = repTemplate;
-
-        ++count;
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString(">> Loaded %u reputation_spillover_template", count);
     sLog.outString();
 }
